refactor event modules
